fix: 修复 flutter analyze 剩余类型错误（CI lint 门禁）
- library_service.dart: box.values 为 List<dynamic>，cast 为 List<Map<dynamic,dynamic>> 后再返回
- zoom_curve_painter.dart / focal_point_overlay.dart: clamp() 返回 num，补 toDouble() 匹配 double 声明/Offset 构造
- home_screen.dart: 导出图标改用确定存在的 upload 系列

+++ lib/features/editor/zoom_curve_painter.dart
@@ -37,8 +37,9 @@ class ZoomCurvePainter extends CustomPainter {
     final curveAreaHeight = trackTop - curveAreaTop;
 
     double yForScale(double scale) {
-      final ratio =
-          ((scale - scaleMin) / (scaleMax - scaleMin)).clamp(0.0, 1.0);
+      final ratio = ((scale - scaleMin) / (scaleMax - scaleMin))
+          .clamp(0.0, 1.0)
+          .toDouble();
       return curveAreaTop + (1.0 - ratio) * curveAreaHeight;
     }
